# sql/motion.openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "motion",
    "description": "",
    "version": "1.0.0"
  },
...
  "paths": {
...
    "/video/publish": {
      "post": {
        "summary": "发布视频(未完成)",
        "x-apifox-folder": "video",
        "x-apifox-status": "released",
        "deprecated": false,
        "description": "开发未完成（98%进度）",
        "tags": [
          "video"
        ],
        "parameters": [
          {
            "name": "description",
            "in": "query",
            "description": "视频描述",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "categoryId",
            "in": "query",
            "description": "分类id",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "tags",
            "in": "query",
            "description": "标签",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "description": "文件",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseResult",
                  "description": "com.ocj.security.commom.ResponseResult"
                },
                "examples": {
                  "1": {
                    "summary": "成功示例",
                    "value": {
                      "code": 0,
                      "msg": "",
                      "data": {}
                    }
                  }
                }
              }
            }
          }
        },
        "x-run-in-apifox": "https://apifox.com/web/project/3485064/apis/api-122980754-run"
      }
    },
    "/video/list": {
      "get": {
        "summary": "获取视频列表",
        "x-apifox-folder": "video",
        "x-apifox-status": "released",
        "deprecated": false,
        "description": "",
        "tags": [
          "video"
        ],
        "parameters": [
          {
            "name": "currentPage",
            "in": "query",
            "description": "分页当前页（必填）",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "description": "分页每页大小（必填）",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "search",
            "in": "query",
            "description": "搜索关键字",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "categoryId",
            "in": "query",
            "description": "类型id筛选",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "tag",
            "in": "query",
            "description": "标签筛选",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseResult",
                  "description": "com.ocj.security.commom.ResponseResult"
                },
                "examples": {
                  "1": {
                    "summary": "成功示例",
                    "value": {
                      "code": 0,
                      "msg": "",
                      "data": {}
                    }
                  }
                }
              }
            }
          }
        },
        "x-run-in-apifox": "https://apifox.com/web/project/3485064/apis/api-122980755-run"
      }
    },
    "/video/one": {
      "get": {
        "summary": "获取每个视频详细信息",
        "x-apifox-folder": "video",
        "x-apifox-status": "released",
        "deprecated": false,
        "description": "",
        "tags": [
          "video"
        ],
        "parameters": [
          {
            "name": "videoId",
            "in": "query",
            "description": "",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseResult",
                  "description": "com.ocj.security.commom.ResponseResult"
                },
                "examples": {
                  "1": {
                    "summary": "成功示例",
                    "value": {
                      "code": 0,
                      "msg": "",
                      "data": {}
                    }
                  }
                }
              }
            }
          }
        },
        "x-run-in-apifox": "https://apifox.com/web/project/3485064/apis/api-122980756-run"
      }
    },
    "/video/{videoId}/comment": {
      "post": {
        "summary": "增加评论",
        "x-apifox-folder": "video",
        "x-apifox-status": "released",
        "deprecated": false,
        "description": "",
        "tags": [
          "video"
        ],
        "parameters": [
          {
            "name": "videoId",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddCommentRequest",
                "description": ""
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseResult%C2%ABAppHttpCodeEnum%C2%BB",
                  "description": ""
                },
                "examples": {
                  "1": {
                    "summary": "成功示例",
                    "value": {
                      "code": 0,
                      "msg": "",
                      "data": ""
                    }
                  }
                }
              }
            }
          }
        },
        "x-run-in-apifox": "https://apifox.com/web/project/3485064/apis/api-122980757-run"
      }
    },
    "/video/{videoId}/comment/list": {
      "get": {
        "summary": "获取评论",
        "x-apifox-folder": "video",
        "x-apifox-status": "released",
        "deprecated": false,
        "description": "",
        "tags": [
          "video"
        ],
        "parameters": [
          {
            "name": "videoId",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseResult%C2%ABCommentDataVO%C2%BB",
                  "description": ""
                },
                "examples": {
                  "1": {
                    "summary": "成功示例",
                    "value": {
                      "code": 0,
                      "msg": "",
                      "data": {
                        "commentCount": 0,
                        "list": [
                          {
                            "id": "",
                            "content": "",
                            "userId": "",
                            "userName": "",
                            "likeCount": 0,
                            "isLiked": false,
                            "avatar": "",
                            "createAt": "",
                            "updateAt": ""
                          }
                        ]
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "x-run-in-apifox": "https://apifox.com/web/project/3485064/apis/api-122980758-run"
      }
    },
    "/video/comment/{commentId}/like": {
      "post": {
        "summary": "点赞评论",
        "x-apifox-folder": "video",
        "x-apifox-status": "released",
        "deprecated": false,
        "description": "",
        "tags": [
          "video"
        ],
        "parameters": [
          {
            "name": "commentId",
            "in": "path",
            "description": "评论id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseResult",
                  "description": "com.ocj.security.commom.ResponseResult"
                },
                "examples": {
                  "1": {
                    "summary": "成功示例",
                    "value": {
                      "code": 0,
                      "msg": "",
                      "data": {}
                    }
                  }
                }
              }
            }
          }
        },
        "x-run-in-apifox": "https://apifox.com/web/project/3485064/apis/api-122980759-run"
      }
    },
...
    "/video/previous/{videoId}": {
      "get": {
        "summary": "当前的视频id查看上一个视频id",
        "x-apifox-folder": "video",
        "x-apifox-status": "released",
        "deprecated": false,
        "description": "",
        "tags": [
          "video"
        ],
        "parameters": [
          {
            "name": "videoId",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseResult",
                  "description": "com.ocj.security.commom.ResponseResult"
                },
                "examples": {
                  "1": {
                    "summary": "成功示例",
                    "value": {
                      "code": 0,
                      "msg": "",
                      "data": {}
                    }
                  }
                }
              }
            }
          }
        },
        "x-run-in-apifox": "https://apifox.com/web/project/3485064/apis/api-122980761-run"
      }
    },
    "/video/next/{videoId}": {
      "get": {
        "summary": "当前的视频id查看下一个视频id",
        "x-apifox-folder": "video",
        "x-apifox-status": "released",
        "deprecated": false,
        "description": "",
        "tags": [
          "video"
        ],
        "parameters": [
          {
            "name": "videoId",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseResult",
                  "description": "com.ocj.security.commom.ResponseResult"
                },
                "examples": {
                  "1": {
                    "summary": "成功示例",
                    "value": {
                      "code": 0,
                      "msg": "",
                      "data": {}
                    }
                  }
                }
              }
            }
          }
        },
        "x-run-in-apifox": "https://apifox.com/web/project/3485064/apis/api-122980762-run"
      }
    }
  },
  "components": {
    "schemas": {
...
      "AddCommentRequest": {
        "type": "object",
        "properties": {
          "content": {
            "type": "string",
            "description": "评论内容",
            "nullable": true
          }
        },
        "x-apifox-orders": [
          "content"
        ],
        "x-apifox-ignore-properties": [],
        "x-apifox-folder": ""
      },
...
      "ResponseResult": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "description": "",
            "nullable": true
          },
          "msg": {
            "type": "string",
            "description": "",
            "nullable": true
          },
          "data": {
            "description": "",
            "type": "object",
            "properties": {}
          }
        },
        "x-apifox-orders": [
          "code",
          "msg",
          "data"
        ],
        "x-apifox-ignore-properties": [],
        "x-apifox-folder": ""
      },
...
    }
  },
  "servers": [
    {
      "url": "127.0.0.1:8060",
      "description": "开发环境"
    }
  ]
}